requirements.txt
Faker==40.1.0
mysql-connector-python==9.5.0
psycopg2-binary==2.9.11
tzdata==2025.3
